// src/app/model/profile.ts
/**
 * Created by boot on 08/01/2018.
 */
/**
 * Created by boot on 21/12/2017.
 */
export class Profile {
  last_name: string;
  first_name: string;
  job_title: string;
  email: string;
  image: string;
  summary: string;
  bird_date: string;
  phone: string;
  contact_email: string;
  facebook: string;
  skype: string;
  twitter: string;
  address: string;
}